backends/fastapi/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
sqlalchemy>=2.0
pyjwt>=2.9
httpx>=0.28
pytest>=8.0
